backend/requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
pytest-asyncio
SQLAlchemy
psycopg2-binary
python-dotenv
joblib
scikit-learn
pandas
python-jose[cryptography]
